error handling for uninstall script

+++ PSGallery/Private/Uninstall-VisualizationSetup.ps1
@@ -21,6 +21,7 @@ function Uninstall-VisualizationSetup {
         Adam Yarborough         1.1             11/07/2018          Integration of Hal's work and updating.
         Adam Yarborough         1.2             12/07/2018          Remove only Grafana, Influx, and NSSM 
                                                                     items from $MonitoringPath
+        Ryan Butler             1.3             24/07/2018          Error and item checking
     .PARAMETER MonitoringPath
         Folder path to download files needed for monitoring process
     .EXAMPLE
@@ -53,29 +54,83 @@ function Uninstall-VisualizationSetup {
             return
         }
         #>
+        
         $NSSMEXE = "$nssm\win64\nssm.exe"
-        #Remove Grafana Service
-        Write-Output "Removing Grafana Server service"
-        & $nssmexe Stop "Grafana Server"
-        & $nssmexe Remove "Grafana Server" confirm
-        #Remove Influx Service
-        Write-Output "Removing InfluxDB Server service"
-        & $nssmexe Stop "InfluxDB Server"
-        & $nssmexe Remove "InfluxDB Server" confirm
+        if (test-path $NSSMEXE)
+        {
+            #Remove Grafana Service
+            Write-Output "Removing Grafana Server service"
+            try {
+                & $nssmexe Stop "Grafana Server"}
+            catch{
+                Write-Warning $($_.Exception.Message)
+            }
+
+            try {
+                & $nssmexe Remove "Grafana Server" confirm
+            }
+            catch{
+                Write-Warning $($_.Exception.Message)
+            }
+
+            #Remove Influx Service
+            Write-Output "Removing InfluxDB Server service"
+            try {
+                & $nssmexe Stop "InfluxDB Server"
+            }
+            catch{
+                Write-Warning $($_.Exception.Message)
+            }
+            
+            try {
+                & $nssmexe Remove "InfluxDB Server" confirm
+            }
+            catch{
+                Write-Warning $($_.Exception.Message)
+            }
+        }
+        else {
+            Write-Warning "NSSM.EXE NOT FOUND. Skipping services."
+        }
 
         #Remove service Directories, all of them.  Scorched earth.
         Write-Output "Removing program directories"
-        Remove-Item -path $Grafana -Recurse 
-        Remove-Item -path $Influx -Recurse 
-        Remove-Item -path $NSSM -Recurse
+        if(-not ([string]::IsNullOrWhiteSpace($Grafana)))
+        {
+            Remove-Item -path $Grafana -Recurse
+        }
+        if(-not ([string]::IsNullOrWhiteSpace($Influx)))
+        {
+            Remove-Item -path $Influx -Recurse
+        }
+        if(-not ([string]::IsNullOrWhiteSpace($NSSM)))
+        {
+            Remove-Item -path $NSSM -Recurse
+        }
 
         #Remove Variable
-        Remove-Item Env:\Home
+        try{
+            Remove-Item Env:\Home -ErrorAction stop
+        }
+        catch{
+            write-warning "Issues removing Influx DB environment variable Home.  Probably already deleted."
+        }
 
         #open FW for Grafana
         Write-Output "Remove Firewall Rules for Grafana and InfluxDB"
-        Remove-NetFirewallRule -DisplayName "Grafana Server"
-        Remove-NetFirewallRule -DisplayName "InfluxDB Server"
+        try {
+            Remove-NetFirewallRule -DisplayName "Grafana Server" -ErrorAction stop
+        }
+        catch{
+            Write-Warning $($_.Exception.Message)
+        }
+        
+        try {
+            Remove-NetFirewallRule -DisplayName "InfluxDB Server" -ErrorAction stop
+        }
+        catch {
+          Write-Warning $($_.Exception.Message)
+        }
     }
 
     end {
